src/requirements.txt
beautifulsoup4==4.12.0
Flask==2.2.3
Pyrebase4==4.6.0
Requests==2.29.0
